requirements-cuepilot.txt
# CuePilot Integration Dependencies
# Additional requirements for the integrated CuePilot functionality

# Core LLM Integration
openai>=1.0.0
PyYAML>=6.0
requests>=2.28.0

# Audio Processing & Transcription
pyaudio>=0.2.11
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Data Processing
python-pptx>=0.6.21

# System Integration (already in main requirements but ensuring versions)
psutil>=5.9.0

# macOS specific (for PowerPoint detection)
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"

# Windows specific (for PowerPoint detection)
pywin32>=306; sys_platform == "win32"

# Optional: Enhanced audio processing
soundfile>=0.12.1
librosa>=0.10.0

# Development & Testing
pytest>=7.4.0